+++ kinto_redis/storage.py
@@ -138,13 +138,9 @@ class Storage(MemoryBasedStorage):
 
     @wrap_redis_error
     def create(self, collection_id, parent_id, record, id_generator=None,
-               unique_fields=None, id_field=DEFAULT_ID_FIELD,
+               id_field=DEFAULT_ID_FIELD,
                modified_field=DEFAULT_MODIFIED_FIELD,
                auth=None):
-        self.check_unicity(collection_id, parent_id, record,
-                           unique_fields=unique_fields, id_field=id_field,
-                           for_creation=True)
-
         record = record.copy()
         id_generator = id_generator or self.id_generator
         _id = record.setdefault(id_field, id_generator())
@@ -187,14 +183,11 @@ class Storage(MemoryBasedStorage):
 
     @wrap_redis_error
     def update(self, collection_id, parent_id, object_id, record,
-               unique_fields=None, id_field=DEFAULT_ID_FIELD,
+               id_field=DEFAULT_ID_FIELD,
                modified_field=DEFAULT_MODIFIED_FIELD,
                auth=None):
         record = record.copy()
         record[id_field] = object_id
-        self.check_unicity(collection_id, parent_id, record,
-                           unique_fields=unique_fields, id_field=id_field)
-
         self.set_record_timestamp(collection_id, parent_id, record,
                                   modified_field=modified_field)